@@ -1,21 +1,22 @@
 package com.baeldung.hibernate.oneToMany.main;
 
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
 import com.baeldung.hibernate.oneToMany.config.HibernateAnnotationUtil;
 import com.baeldung.hibernate.oneToMany.model.Cart;
 import com.baeldung.hibernate.oneToMany.model.Item;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 
 public class HibernateOneToManyAnnotationMainIntegrationTest {
 
@@ -23,12 +24,12 @@ public class HibernateOneToManyAnnotationMainIntegrationTest {
 
 	private Session session;
 
-	@BeforeClass
+	@BeforeAll
 	public static void beforeTests() {
 		sessionFactory = HibernateAnnotationUtil.getSessionFactory();
 	}
 
-	@Before
+	@BeforeEach
 	public void setUp() {
 		session = sessionFactory.openSession();
 		session.beginTransaction();
@@ -36,8 +37,8 @@ public class HibernateOneToManyAnnotationMainIntegrationTest {
 
 	@Test
 	public void givenSession_checkIfDatabaseIsEmpty() {
-		Cart cart = session.get(Cart.class, 1L);
-		assertNull(cart);
+		Cart cart = session.find(Cart.class, 1L);
+		Assertions.assertNull(cart);
 	}
 
 	@Test
@@ -45,7 +46,7 @@ public class HibernateOneToManyAnnotationMainIntegrationTest {
 		Cart cart = new Cart();
 		Set<Item> cartItems = cart.getItems();
 
-		assertNull(cartItems);
+		Assertions.assertNull(cartItems);
 
 		Item item1 = new Item();
 		item1.setCart(cart);
@@ -66,18 +67,18 @@ public class HibernateOneToManyAnnotationMainIntegrationTest {
 
 		session = sessionFactory.openSession();
 		session.beginTransaction();
-		cart = session.get(Cart.class, 1L);
+		cart = session.find(Cart.class, 1L);
 
 		assertNotNull(cart);
 	}
 
-	@After
+	@AfterEach
 	public void tearDown() {
 		session.getTransaction().commit();
 		session.close();
 	}
 
-	@AfterClass
+	@AfterAll
 	public static void afterTests() {
 		sessionFactory.close();
 	}
