@@ -1,7 +1,6 @@
 package com.baeldung.hibernate.customtypes;
 
 import org.hibernate.HibernateException;
-import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.metamodel.spi.ValueAccess;
 import org.hibernate.usertype.CompositeUserType;
 
@@ -20,13 +19,14 @@ public class PhoneNumberType implements CompositeUserType<PhoneNumber> {
             case 2:
                 return component.getNumber();
             default:
-                throw new IllegalArgumentException(property + " is an invalid property index for class type " + component.getClass().getName());
+                throw new IllegalArgumentException(property + " is an invalid property index for class type " + component.getClass()
+                    .getName());
         }
     }
 
     @Override
-    public PhoneNumber instantiate(ValueAccess values, SessionFactoryImplementor sessionFactory) {
-        return new PhoneNumber(values.getValue(0, Integer.class), values.getValue(1, Integer.class), values.getValue(2,Integer.class));
+    public PhoneNumber instantiate(ValueAccess values) {
+        return new PhoneNumber(values.getValue(0, Integer.class), values.getValue(1, Integer.class), values.getValue(2, Integer.class));
     }
 
     @Override
@@ -41,10 +41,12 @@ public class PhoneNumberType implements CompositeUserType<PhoneNumber> {
 
     @Override
     public boolean equals(PhoneNumber x, PhoneNumber y) {
-        if (x == y)
+        if (x == y) {
             return true;
-        if (Objects.isNull(x) || Objects.isNull(y))
+        }
+        if (Objects.isNull(x) || Objects.isNull(y)) {
             return false;
+        }
 
         return x.equals(y);
     }
@@ -56,8 +58,9 @@ public class PhoneNumberType implements CompositeUserType<PhoneNumber> {
 
     @Override
     public PhoneNumber deepCopy(PhoneNumber value) {
-        if (Objects.isNull(value))
+        if (Objects.isNull(value)) {
             return null;
+        }
 
         return new PhoneNumber(value.getCountryCode(), value.getCityCode(), value.getNumber());
     }
