@@ -7,7 +7,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.retry.support.RetrySynchronizationManager;
 import org.springframework.stereotype.Service;
 import org.apache.commons.lang3.StringUtils;
-
+import org.springframework.resilience.annotation.ConcurrencyLimit;
 @Service
 public class MyServiceImpl implements MyService {
 
@@ -15,9 +15,11 @@ public class MyServiceImpl implements MyService {
 
     @Override
     public void retryService() {
+        // This is correct for logging the retry count (0 on first attempt, 1 on first retry, etc.)
         logger.info("Retry Number: " + RetrySynchronizationManager.getContext()
             .getRetryCount());
         logger.info("throw RuntimeException in method retryService()");
+        // Always throws the exception to force the retry aspect to kick in
         throw new RuntimeException();
     }
 
@@ -33,6 +35,7 @@ public class MyServiceImpl implements MyService {
     public void retryServiceWithCustomization(String sql) throws SQLException {
         if (StringUtils.isEmpty(sql)) {
             logger.info("throw SQLException in method retryServiceWithCustomization()");
+            // Correctly throws SQLException to trigger the 2 retry attempts
             throw new SQLException();
         }
     }
@@ -54,5 +57,20 @@ public class MyServiceImpl implements MyService {
     public void templateRetryService() {
         logger.info("throw RuntimeException in method templateRetryService()");
         throw new RuntimeException();
+    }
+
+    // **NEW Implementation for Concurrency Limit**
+    @Override
+    @ConcurrencyLimit(5)
+    public void concurrentLimitService() {
+        // The ConcurrencyLimit aspect will wrap this method.
+        logger.info("Concurrency Limit Active. Current Thread: " + Thread.currentThread().getName());
+        // Simulate a time-consuming task to observe throttling
+        try {
+            Thread.sleep(1000); // Correctly blocks to hold the lock
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+        }
+        logger.info("Concurrency Limit Released. Current Thread: " + Thread.currentThread().getName());
     }
 }
