@@ -2,7 +2,6 @@ package com.baeldung.async;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,18 +18,22 @@ public class AsyncServiceUnitTest {
     @Autowired
     private AsyncService asyncServiceExample;
 
-    // tests
-
     @Test
     public void testAsyncAnnotationForMergedServicesResponse() throws InterruptedException, ExecutionException {
+         
+        //   we use the injected bean 'asyncServiceExample'.
         CompletableFuture<String> completableFuture = asyncServiceExample.asyncMergeServicesResponse();
 
+        System.out.println("Invoking asynchronous methods. " + Thread.currentThread().getName());
+
+        // This loop simulates the calling thread continuing its work while waiting for the async result.
         while (true) {
             if (completableFuture.isDone()) {
+                System.out.println("Result from asynchronous process - " + completableFuture.get());
                 break;
             }
+            System.out.println("Continue doing something else. ");
             Thread.sleep(1000);
         }
     }
-
 }
