@@ -13,12 +13,13 @@ public class AsyncWithXMLIntegrationTest {
     @Autowired
     private AsyncComponent asyncAnnotationExample;
 
-    // tests
-
     @Test
     public void testAsyncAnnotationForMethodsWithVoidReturnType() throws InterruptedException {
+        // Invoking the async method. The calling thread will immediately proceed.
         asyncAnnotationExample.asyncMethodWithVoidReturnType();
+        
+        // Sleep is required in void methods to ensure the asynchronous thread 
+        // has time to execute before the test completes and the application context shuts down.
         Thread.sleep(2000);
     }
-
 }
